+++ src/main/java/tarawrr/DeleteCommand.java
@@ -1,3 +1,4 @@
+package tarawrr;
 public class DeleteCommand extends Command {
     private int index;
