+++ src/main/java/tarawrr/Event.java
@@ -1,3 +1,4 @@
+package tarawrr;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
