+++ src/main/java/tarawrr/HelperFunctions.java
@@ -1,3 +1,4 @@
+package tarawrr;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
