+++ src/main/java/tarawrr/MarkCommand.java
@@ -1,4 +1,4 @@
-import java.util.ArrayList;
+package tarawrr;
 
 public class MarkCommand extends Command {
     private int index;
