+++ src/main/java/tarawrr/Parser.java
@@ -1,3 +1,4 @@
+package tarawrr;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
