+++ src/main/java/tarawrr/Storage.java
@@ -1,3 +1,4 @@
+package tarawrr;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
