+++ src/main/java/tarawrr/TarawrrException.java
@@ -1,3 +1,4 @@
+package tarawrr;
 public class TarawrrException extends Exception {
     public TarawrrException(String message) {
         super(message);
