+++ src/main/java/tarawrr/Task.java
@@ -1,3 +1,4 @@
+package tarawrr;
 public abstract class Task {
     private boolean status;
     private String name;
